+++ document-store/src/integrationTest/resources/query/function_expression_group_by_response.json
